Find and remove extra player files.
There are player files for player names that do not appear in the auth.txt
file. These are probably left over from a previous glitch in player
pruning so it should be safe to remove all of them. Each file is moved to
the pruned directory and the date of removal is logged in extra.pruned
just in case.